Documents/maktab_khoone_advance/requirements.txt
django==4.2
python-decouple
Pillow
djangorestframework
markdown
django-filter